分布式ID +

+++ src/main/resources/mysql/dict_order_type.sql
@@ -1,6 +1,6 @@
 #订单状态字典表
 CREATE TABLE dict_order_type(
-id INT AUTO_INCREMENT,
+id BIGINT AUTO_INCREMENT,
 order_type VARCHAR(200),
 PRIMARY KEY(id)
 );
